fix(ios): narrow CC_SHA256_DIGEST_LENGTH to Int for ByteArray ctor
Kotlin/Native 2.3.x platform-lib bindings for CommonCrypto expose
CC_SHA256_DIGEST_LENGTH as Long, but ByteArray(size: Int) needs an Int.
Without the narrowing the iosArm64 compile fails with a type mismatch
on the digest buffer allocation, blocking iOS Test Target Compile on
#374.

+++ shared/src/iosMain/kotlin/com/devil/phoenixproject/data/auth/OAuth.ios.kt
@@ -41,7 +41,9 @@ internal actual fun generateSecureRandomBytes(size: Int): ByteArray {
 
 @OptIn(ExperimentalForeignApi::class)
 internal actual fun sha256(input: ByteArray): ByteArray {
-    val digest = ByteArray(CC_SHA256_DIGEST_LENGTH)
+    // CC_SHA256_DIGEST_LENGTH comes from CommonCrypto as Long in Kotlin/Native
+    // 2.3.x platform libs; ByteArray(size: Int) needs the explicit narrowing.
+    val digest = ByteArray(CC_SHA256_DIGEST_LENGTH.toInt())
     if (input.isEmpty()) {
         digest.usePinned { digestPinned ->
             CC_SHA256(null, 0u, digestPinned.addressOf(0).reinterpret())
